+++ tests/test_timerit.py
@@ -111,6 +111,37 @@ def test_hacked_timerit_verbose():
             time per loop: best=42.000 s, mean=42.000 +- 0.0 s
         ''').strip()
 
+    with CaptureStdout(suppress=False) as cap:
+        HackedTimerit(num=None, label='foo', verbose=0, min_duration=100).call(lambda: None)
+    assert cap.text.strip() == textwrap.dedent(
+        '''
+        ''').strip()
+
+    with CaptureStdout(suppress=False) as cap:
+        HackedTimerit(num=None, label='foo', verbose=1, min_duration=100).call(lambda: None)
+    assert cap.text.strip() == textwrap.dedent(
+        '''
+        Timed best=42.000 s, mean=42.000 +- 0.0 s for foo
+        ''').strip()
+
+    with CaptureStdout(suppress=False) as cap:
+        HackedTimerit(num=None, label='foo', verbose=2, min_duration=100).call(lambda: None)
+    assert cap.text.strip() == textwrap.dedent(
+        '''
+        Timed foo for: 3 loops, best of 3
+            time per loop: best=42.000 s, mean=42.000 +- 0.0 s
+        ''').strip()
+
+    with CaptureStdout(suppress=False) as cap:
+        HackedTimerit(num=None, label='foo', verbose=3, min_duration=100).call(lambda: None)
+    assert cap.text.strip() == textwrap.dedent(
+        '''
+        Timing foo for: 100.000s
+        Timed foo for: 3 loops, best of 3
+            body took: 126.000 s
+            time per loop: best=42.000 s, mean=42.000 +- 0.0 s
+        ''').strip()
+
 
 def test_timer_default_verbosity():