+++ timerit/core.py
@@ -279,11 +279,12 @@ class Timerit:
     _default_precision_type = 'f'  # could also be reasonably be 'g' or ''
 
     def __init__(self, num=1, label=None, bestof=3, unit=None, verbose=None,
-                 disable_gc=True, timer_cls=None):
+                 disable_gc=True, timer_cls=None, min_duration=0.2):
         """
         Args:
-            num (int):
-                Number of times to run the loop. Defaults to 1.
+            num (int | None):
+                Number of times to run the loop.  If None, then the number of
+                loops is determined by ``min_duration``.
 
             label (str | None):
                 An identifier for printing and differentiating between
@@ -312,6 +313,10 @@ class Timerit:
             timer_cls (None | Any):
                 If specified, replaces the default :class:`Timer` class with a
                 customized one. Mainly useful for testing.
+
+            min_duration (float):
+                Run the loop until the given amount of time has elapsed.
+                Ignored unless ``num is None``.
         """
         if verbose is None:
             verbose = bool(label)
@@ -321,6 +326,7 @@ class Timerit:
         self.bestof = bestof
         self.unit = unit
         self.verbose = verbose
+        self.min_duration = min_duration
 
         self.times = []
         self.total_time = 0
@@ -416,21 +422,28 @@ class Timerit:
 
     def __iter__(self):
         if self.verbose >= 3:
-            print(self._status_line(tense='present'))
+            print(self._status_line())
 
         self.n_loops = 0
         self.total_time = 0
 
-        # TODO: if num is unspecified, can we just run for a set amount of time
-        # and then stop?
-
         bg_timer = self._bg_timer
         fg_timer = self._fg_timer
+
         # disable the garbage collector while timing
         with _SetGCState(enable=False), _SetDisplayHook():
             # Core timing loop
-            for _ in it.repeat(None, self.num):
-                # Start background timer (in case the user doesn't use fg_timer)
+            while True:
+
+                if self.num is None:
+                    if self.total_time > self.min_duration:
+                        break
+                else:
+                    if self.n_loops >= self.num:
+                        break
+
+                # Start background timer (in case the user doesn't use
+                # fg_timer)
                 # Yield foreground timer to let the user run a block of code
                 # When we return from yield the user code will have just finished
                 # Then record background time + loop overhead
@@ -447,7 +460,7 @@ class Timerit:
                 self.total_time += block_time
                 self.n_loops += 1
         # Timing complete, print results
-        if len(self.times) != self.num:
+        if self.num and len(self.times) != self.num:
             raise AssertionError(
                 'incorrectly recorded times, need to reset timerit object')
 
@@ -682,27 +695,44 @@ class Timerit:
                                  pr1=pr1, pr2=pr2)
         return unit_str
 
-    def _status_line(self, tense='past'):
+    def _status_line(self):
         """
         Text indicating what has been / is being done.
-
-        Args:
-            tense (str): Either 'past' or 'present'
 
         Returns:
             str:
 
         Example:
-            >>> from timerit import Timer
-            >>> print(Timerit()._status_line(tense='past'))
-            Timed for: 1 loops, best of 1
-            >>> print(Timerit()._status_line(tense='present'))
+            >>> from timerit import Timerit
+            >>> from time import sleep
+            >>> t = Timerit()
+            >>> print(t._status_line())
             Timing for: 1 loops, best of 1
+            >>> t.call(sleep, 0.01)
+            >>> print(t._status_line())
+            Timed for: 1 loops, best of 1
         """
+
+        tense = 'present' if self.n_loops is None else 'past'
+
         action = {'past': 'Timed',  'present': 'Timing'}[tense]
-        line = '{action} {label}for: {num:d} loops, best of {bestof:d}'.format(
+
+        if tense == 'present':
+            if self.num is None:
+                duration = '{:.3f}s'.format(self.min_duration)
+            else:
+                duration = '{num:d} loops, best of {bestof:d}'.format(
+                        num=self.num,
+                        bestof=min(self.bestof, self.num))
+        else:
+            duration = '{num:d} loops, best of {bestof:d}'.format(
+                    num=self.n_loops,
+                    bestof=min(self.bestof, self.n_loops or 1))
+
+        line = '{action} {label}for: {duration}'.format(
             label=self.label + ' ' if self.label else '',
-            action=action, num=self.num, bestof=min(self.bestof, self.num))
+            action=action, duration=duration)
+
         return line
 
     def summary(self, stat='mean'):
@@ -775,7 +805,7 @@ class Timerit:
         lines = []
         if verbose >= 2:
             # use a multi-line format for high verbosity
-            lines.append(self._status_line(tense='past'))
+            lines.append(self._status_line())
             if verbose >= 3:
                 unit, mag = _choose_unit(self.total_time, self.unit,
                                          self._asciimode)
@@ -858,12 +888,13 @@ class _SetGCState(object):
         else:
             gc.disable()
 
+
 class _SetDisplayHook(object):
     """
-    Context manager to prevent the REPL interpreter from printing the value of 
+    Context manager to prevent the REPL interpreter from printing the value of
     any expressions within the for loop that don't evaluate to None.
 
-    Printing is relatively expensive, and so this behavior can easily lead to 
+    Printing is relatively expensive, and so this behavior can easily lead to
     timings that are much longer than they should be.
     """
     def __enter__(self):
@@ -872,8 +903,6 @@ class _SetDisplayHook(object):
 
     def __exit__(self, ex_type, ex_value, trace):
         sys.displayhook = self._orig_display_hook
-
-
 
 
 def _chunks(seq, size):